% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio.R
\name{db_con}
\alias{db_con}
\title{Create a MariaDB Connection}
\usage{
db_con(.cnf = NULL, group = NULL)
}
\arguments{
\item{.cnf}{Path to a MariaDB option file. If \code{NULL}, resolved from
\code{cnf_path} or \code{GPXUI_CNF}.}

\item{group}{Option group in \code{.cnf}. If \code{NULL}, resolved from global
\code{group} or \code{"gpxui"}.}
}
\value{
A MariaDB connection.
}
\description{
Creates a MariaDB connection from an option file.
}
\details{
If \code{.cnf} is \code{NULL}, \code{db_con()} first looks for \code{cnf_path} in the global
environment, then for the \code{GPXUI_CNF} environment variable.

If \code{group} is \code{NULL}, \code{db_con()} first looks for \code{group} in the global
environment, then uses \code{"gpxui"}.

The database/schema should be configured in the selected option group.
}
