% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio.R
\name{db_get}
\alias{db_get}
\title{Run a MariaDB Query and Return a Data Table}
\usage{
db_get(query, ..., params = NULL)
}
\arguments{
\item{query}{SQL query passed to \code{\link[DBI:dbGetQuery]{DBI::dbGetQuery()}}.}

\item{...}{Additional arguments passed to \code{\link[=db_con]{db_con()}}.}

\item{params}{Optional query parameters passed to \code{\link[DBI:dbGetQuery]{DBI::dbGetQuery()}}.}
}
\value{
A \code{data.table}.
}
\description{
Run a MariaDB Query and Return a Data Table
}
