% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpsid.R
\name{garminID}
\alias{garminID}
\title{read_gpsID}
\usage{
garminID(p)
}
\arguments{
\item{p}{path to startup.txt}
}
\description{
read_gpsID
}
\examples{
p <- system.file(package = "gpxui", "Garmin65s", "Garmin", "startup.txt")
}
