% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_modules.R
\name{gpx_file_upload_check}
\alias{gpx_file_upload_check}
\title{gpx_file_upload_check}
\usage{
gpx_file_upload_check(x)
}
\arguments{
\item{x}{a data.frame returned by UI after dir upload. See \code{\link[=dirInput]{dirInput()}}}
}
\value{
A Shiny tag object describing the selected directory.
}
\description{
gpx_file_upload_check
}
