% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{gpx_server}
\alias{gpx_server}
\title{GPX Shiny Server}
\usage{
gpx_server(.cnf = NULL, group = NULL)
}
\arguments{
\item{.cnf}{Path to a MariaDB option file passed to \code{\link[=db_con]{db_con()}}.}

\item{group}{Option group in \code{.cnf} passed to \code{\link[=db_con]{db_con()}}.}
}
\value{
A Shiny server function.
}
\description{
GPX Shiny Server
}
