% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{gpx_ui}
\alias{gpx_ui}
\title{GPX Shiny UI}
\usage{
gpx_ui(gps_ids = 1:10, export_tables = c("mid_points", "mid_tracks"))
}
\arguments{
\item{gps_ids}{GPS ids shown in the explorer filter.}

\item{export_tables}{Database tables/views available for export.}
}
\value{
A Shiny UI object.
}
\description{
GPX Shiny UI
}
