% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx_read.R
\name{read_all_waypoints}
\alias{read_all_waypoints}
\title{read_all_waypoints}
\usage{
read_all_waypoints(ff, int_names_only = TRUE, gpsid)
}
\arguments{
\item{ff}{a vector of file names}

\item{int_names_only}{keep only numeric names}

\item{gpsid}{the id of the gps.}
}
\description{
read_all_waypoints
}
\examples{
ff = list.files(system.file(package = "gpxui", "Garmin65s"), full.names = TRUE, recursive = TRUE)
read_all_waypoints(ff, gpsid = 1)
}
