% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx_read.R
\name{read_gpx}
\alias{read_gpx}
\alias{read_waypoints}
\alias{read_tracks}
\title{Read Waypoints and Tracks}
\usage{
read_waypoints(x)

read_tracks(x)
}
\arguments{
\item{x}{A path to a GPX file.}
}
\value{
A \code{data.table} containing:
\describe{
\item{gps_point}{For waypoints: waypoint name. For tracks: track segment identifier.}
\item{datetime_}{Timestamp of the recorded point (if available).}
\item{ele}{Elevation in meters.}
\item{lon}{Longitude.}
\item{lat}{Latitude.}
}
}
\description{
These functions are wrappers around \code{sf::st_read}, reading waypoints and tracks from GPX files.
If the file is corrupt or empty, the functions return an empty \code{data.table} with a warning.
}
\examples{
f1 = system.file(package = "gpxui", "Garmin65s", "GPX", "Waypoints_20-APR-23.gpx")
f2 = system.file(package = "gpxui", "Garmin65s", "GPX", "Waypoints_empty.gpx")
f3 = system.file(package = "gpxui", "Garmin65s", "GPX", "Current", "Current.gpx")
read_waypoints(f1)
read_waypoints(f2)
read_tracks(f1)
read_tracks(f2)
read_tracks(f3)

}
