% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gpx.R
\name{read_tracks}
\alias{read_tracks}
\title{read_tracks}
\usage{
read_tracks(x)
}
\description{
read_tracks
}
\examples{
f = system.file(package = "gpxui", "Garmin65s","Garmin", "GPX", "Current", "Current.gpx")
read_tracks(f)
}
