% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{st_bbox_all}
\alias{st_bbox_all}
\title{Combine Bounding Boxes}
\usage{
st_bbox_all(x)
}
\arguments{
\item{x}{A list that may contain \code{sf} objects.}
}
\value{
An \code{sf} \code{bbox}.
}
\description{
Return one bounding box for all non-empty \code{sf} objects in a list. Empty or
non-\code{sf} inputs are ignored.
}
