% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI_csvExport.R
\name{ui_csvExport}
\alias{ui_csvExport}
\title{ui_csvExport}
\usage{
ui_csvExport(...)
}
\arguments{
\item{...}{aguments passed to \code{\link[bslib:bs_theme]{bslib::bs_theme()}}}
}
\description{
ui_csvExport
}
