% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx_write.R
\name{waypoints_to_db}
\alias{waypoints_to_db}
\title{waypoints_to_db}
\usage{
waypoints_to_db(x)
}
\description{
waypoints_to_db
}
\examples{
x <- read_all_waypoints(system.file(package = "gpxui", "Garmin65s"))
}
