% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_dirInput_output}
\alias{as_dirInput_output}
\title{as_dirInput_output}
\usage{
as_dirInput_output(dr)
}
\description{
as_dirInput_output
}
\examples{
system.file(package = "gpxui", "Garmin65s") |>
  as_dirInput_output()
}
