% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx_read.R
\name{deviceID}
\alias{deviceID}
\title{deviceID}
\usage{
deviceID(p)
}
\arguments{
\item{x}{a data.frame  uploaded to the server by dirInput}
}
\description{
deviceID
}
\examples{
p <- system.file(package = "gpxui", "Garmin65s", "GPX") |> as_dirInput_output()
deviceID(p)
}
