% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deviceID}
\alias{deviceID}
\title{deviceID}
\usage{
deviceID(p)
}
\arguments{
\item{x}{path to where the device ID is stored}
}
\description{
deviceID
}
\examples{
p <- system.file(package = "gpxui", "Garmin65s", "GPX", "DEVICE_ID.txt")
deviceID(p)
}
