% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_modules.R
\name{dirInput}
\alias{dirInput}
\title{dirInput}
\usage{
dirInput(
  inputId,
  label = NULL,
  buttonLabel = "GPS upload",
  placeholder = "./GARMIN/Garmin/GPX"
)
}
\description{
dirInput
}
\examples{
require(shiny)
# options(shiny.maxRequestSize = 10 * 1024^2)

ui <- fluidPage(
  sidebarLayout(
    sidebarPanel(
      dirInput("fileIn", "Choose dir")
    ),
    mainPanel(
      tableOutput("contents")
    )
  )
)
server <- function(input, output) {
  output$contents <- renderTable({
print(input$fileIn)
if(!is.null(input$fileIn))
data.table(input$fileIn)[, file.size(datapath), by = name]
  })
}
shinyApp(ui, server)
}
