% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx_write.R
\name{DT2gpx}
\alias{DT2gpx}
\title{DT2gpx}
\usage{
DT2gpx(
  x,
  longit = "lon",
  latit = "lat",
  nam,
  symbol = "Bird",
  dest = tempfile(fileext = ".gpx")
)
}
\description{
DT2gpx
}
\examples{
x = system.file(package = "gpxui", "Garmin65s") |> as_dirInput_output() |>
  read_all_waypoints()
outf = tempfile(fileext = ".gpx")
DT2gpx(x, nam = "gps_point", dest = outf)
}
