% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio.R
\name{gpx_to_database}
\alias{gpx_to_database}
\title{gpx_to_database
saves new entries to database}
\usage{
gpx_to_database(con, x, tab)
}
\arguments{
\item{con}{a connection to db}

\item{x}{a data.table; output of read_all_waypoints() or read_all_tracks().}

\item{tab}{database table}
}
\value{
a data.frame containing last entry in db prior to database update and the number of updated entries.
}
\description{
gpx_to_database
saves new entries to database
}
\examples{
require(dbo)
ff <- system.file(package = "gpxui", "Garmin65s") |> as_dirInput_output()
con <- dbcon(server = "localhost", db = "tests")

x <- read_all_waypoints(ff)
gpx_to_database(con, x, tab = "GPS_POINTS")

x <- read_all_tracks(ff)
gpx_to_database(con, x, tab = "GPS_TRACKS")

DBI::dbDisconnect(con)

}
