% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio.R
\name{gpx_to_database}
\alias{gpx_to_database}
\title{gpx_to_database
saves new entries to database}
\usage{
gpx_to_database(con, x, tab)
}
\arguments{
\item{con}{a connection to db}

\item{x}{a data.table; output of read_all_waypoints() or read_all_tracks().}

\item{tab}{database table (GPS_TRACKS, GPS_POINTS)}
}
\value{
a data.frame containing last entry in db prior to database update and the number of updated entries.
}
\description{
gpx_to_database
saves new entries to database
}
