% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio.R
\name{gpx_to_database}
\alias{gpx_to_database}
\title{gpx_to_database
saves new entries to database}
\usage{
gpx_to_database(server, db, x, tab)
}
\arguments{
\item{server}{pass to \code{\link[dbo:dbcon]{dbo::dbcon()}}}

\item{db}{pass to \code{\link[dbo:dbcon]{dbo::dbcon()}}}

\item{x}{output of \code{\link[=read_all_waypoints]{read_all_waypoints()}} or \code{\link[=read_all_tracks]{read_all_tracks()}}.}

\item{tab}{database table (GPS_TRACKS, GPS_POINTS)}
}
\value{
a data.frame containing last entry in db prior to database update and the number of updated entries.
}
\description{
gpx_to_database
saves new entries to database
}
