% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{points_summary}
\alias{points_summary}
\title{points_summary #TODO}
\usage{
points_summary(x)
}
\description{
points_summary #TODO
}
\examples{
x = 
system.file(package = "gpxui", "Garmin65s") |>
as_dirInput_output() |>
read_all_tracks() |>
points_summary()
}
