% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{points_summary}
\alias{points_summary}
\title{points_summary}
\usage{
points_summary(x)
}
\description{
points_summary
}
\examples{
ff <- list.files(system.file(package = "gpxui", "Garmin65s"), full.names = TRUE, recursive = TRUE)
read_all_waypoints(ff, gpsid = 1) |> points_summary()
}
