% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx_read.R
\name{read_all_tracks}
\alias{read_all_tracks}
\title{read_all_tracks}
\usage{
read_all_tracks(ff, gpsid)
}
\arguments{
\item{ff}{a vector of file names}

\item{gpsid}{the id of the gps.}
}
\description{
read_all_tracks
}
\examples{
ff = list.files(system.file(package = "gpxui", "Garmin65s"), full.names = TRUE, recursive = TRUE)
read_all_tracks(ff, gpsid = 1)
}
