% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx_read.R
\name{read_all_tracks}
\alias{read_all_tracks}
\title{read_all_tracks}
\usage{
read_all_tracks(ff)
}
\arguments{
\item{ff}{a data.frame  uploaded to the server by dirInput}
}
\description{
read all tracks from the GPX directory and the gps id from DEVICE_ID.txt when it exists
}
