% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gpx.R
\name{read_all_waypoints}
\alias{read_all_waypoints}
\title{read_all_waypoints}
\usage{
read_all_waypoints(d, int_names_only = TRUE)
}
\description{
read_all_waypoints
}
\examples{
read_all_waypoints(system.file(package = "gpxui", "Garmin65s"))
}
