% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx_read.R
\name{read_all_waypoints}
\alias{read_all_waypoints}
\title{read_all_waypoints}
\usage{
read_all_waypoints(ff, int_names_only = TRUE)
}
\arguments{
\item{ff}{a data.frame  uploaded to the server by dirInput or
a directory path containing gpx files.}

\item{int_names_only}{keep only numeric names}
}
\description{
read all waypoints from the GPX directory and the gps id from DEVICE_ID.txt when it exists
}
\examples{
g = system.file("Garmin65s", package = "gpxui") |> read_all_waypoints()
}
