% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gpx.R
\name{read_all_waypoints}
\alias{read_all_waypoints}
\title{read_all_waypoints}
\usage{
read_all_waypoints(ff, int_names_only = TRUE, sf = FALSE)
}
\arguments{
\item{ff}{a vector of file names}

\item{int_names_only}{keep only numeric names}

\item{sf}{output as a sf dataframe.}
}
\description{
read_all_waypoints
}
\examples{
ff = list.files(system.file(package = "gpxui", "Garmin65s"), full.names = TRUE, recursive = TRUE)
read_all_waypoints(ff)
read_all_waypoints(ff, sf = TRUE)
}
