% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbio.R
\name{read_GPX_table}
\alias{read_GPX_table}
\title{read_GPX_table
Fetch database tables}
\usage{
read_GPX_table(con, tab, dt = "1900-01-01", sf = FALSE)
}
\arguments{
\item{con}{a connection to db}

\item{tab}{database table   (GPS_TRACKS, GPS_POINTS)}

\item{dt}{database valid datetime. Only entries after this are returned. Detault to "1900-01-01"}

\item{sf}{when TRUE returns a sf df. default to FALSE}
}
\description{
read_GPX_table
Fetch database tables
}
