% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx_read.R
\name{read_waypoints}
\alias{read_waypoints}
\title{read_waypoints}
\usage{
read_waypoints(x)
}
\description{
read_waypoints
}
\examples{
f = system.file(package = "gpxui", "Garmin65s","Garmin", "GPX", "Waypoints_20-APR-23.gpx")
read_waypoints(f)
}
