% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{st_bbox_all}
\alias{st_bbox_all}
\title{st_bbox_all
st_bbox on a list}
\usage{
st_bbox_all(x)
}
\arguments{
\item{x}{a list of sf objects, any non-sf objects in the list are silently ignored.}
}
\value{
a st_bbox object or NULL when it fails.
}
\description{
st_bbox_all
st_bbox on a list
}
\examples{
ff <- list.files(system.file(package = "gpxui", "Garmin65s"), full.names = TRUE, recursive = TRUE)
pts = read_all_waypoints(ff, sf = TRUE)
trk = read_all_tracks(ff, sf = TRUE)
st_bbox_all(list(pts, trk))
st_bbox_all(list(pts, NULL))
st_bbox_all(list(1, NULL))
}
