% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{track_summary}
\alias{track_summary}
\title{track_summary}
\usage{
track_summary(x)
}
\description{
track_summary
}
\examples{
ff <- list.files(system.file(package = "gpxui", "Garmin65s"), full.names = TRUE, recursive = TRUE)
read_all_tracks(ff, gpsid = 1) |> track_summary()
}
