% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{track_summary}
\alias{track_summary}
\title{track_summary #TODO}
\usage{
track_summary(x)
}
\description{
track_summary #TODO
}
\examples{
system.file(package = "gpxui", "Garmin65s") |> 
as_dirInput_output() |>
read_all_tracks() |> 
track_summary()
}
